requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
crispy-bootstrap4==2022.1
Django==4.1.4
django-ckeditor==6.5.1
django-clearcache==1.2.1
django-crispy-forms==2.0
django-dotenv==1.4.2
django-js-asset==2.0.0
Pillow==9.4.0
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
gunicorn
requests
django-dotenv
django-storages
boto3